% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{corners_cpp}
\alias{corners_cpp}
\title{Corner coordinates}
\usage{
corners_cpp(x, y)
}
\arguments{
\item{x}{coordinates}

\item{y}{coordinates}
}
\value{
character vector of layer names
}
\description{
Corner coordinates
}
