% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{llh2xyz}
\alias{llh2xyz}
\title{Angular coordinates to X, Y, Z.}
\usage{
llh2xyz(lonlatheight, rad = 6378137, exag = 1)
}
\arguments{
\item{lonlatheight}{matrix or data.frame of lon,lat,height values}

\item{rad}{radius of sphere}

\item{exag}{exaggeration to apply to height values (added to radius)}
}
\value{
matrix
}
\description{
Angular coordinates to X, Y, Z.
}
