% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mesh-plot.R
\name{mesh_plot}
\alias{mesh_plot}
\alias{mesh_plot}
\alias{mesh_plot.BasicRaster}
\alias{mesh_plot}
\alias{mesh_plot.RasterLayer}
\title{Plot as a mesh}
\usage{
mesh_plot(x, crs = NULL, colfun = NULL, add = FALSE, ...)

\method{mesh_plot}{BasicRaster}(x, crs = NULL, colfun = NULL, add = FALSE,
  ...)

\method{mesh_plot}{RasterLayer}(x, crs = NULL, colfun = NULL, add = FALSE,
  ...)
}
\arguments{
\item{x}{object to convert to mesh and plot}

\item{crs}{target map projection}

\item{colfun}{colour function to use, `viridis` is the default}

\item{add}{add to existing plot or start a new one}

\item{...}{ignored}
}
\value{
nothing
}
\description{
Plot as a mesh
}
\examples{

plot(wrld_simpl)
mesh_plot(worldll)

mesh_plot(worldll, crs = "+proj=laea")
mesh_plot(worldll, crs = "+proj=moll")
prj <- "+proj=lcc +datum=WGS84 +lon_0=147 +lat_0=-40 +lat_1=-55 +lat_2=-20"
plot(spTransform(subset(wrld_simpl, NAME == "Australia"), prj))
## this is near-lossless transformation, at native resolution
mesh_plot(etopo, crs = prj, add = TRUE, colfun = function(n = 20) grey(seq(0, 1, length = n)))
mesh_plot(worldll, crs = prj, add = TRUE)
}
