% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qsc.R
\name{qsc}
\alias{qsc}
\title{Quadrilateralized Spherical Cube (QSC)}
\usage{
qsc()
}
\value{
mesh3d
}
\description{
The QSC is a set of six equal area projections for each side of the cube. Here
a raw rendition of the cube is returned as six quad primitives in a \code{mesh3d} object.
}
\details{
It's not clear if this is useful.
}
\examples{
library(rgl); #rgl.clear()
wire3d(qsc())
if ( rgl::rgl.useNULL()) rglwidget()
}
\references{
\url{https://github.com/OSGeo/proj.4/wiki/Qsc}
}
