% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quadmesh-package.R
\docType{package}
\name{quadmesh-package}
\alias{quadmesh}
\alias{quadmesh-package}
\title{quadmesh: Quadrangle Mesh}
\description{
Create a standard quadrangle mesh from gridded data ('rasters') for flexible
plotting and conversion to other mesh forms.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/mdsumner/quadmesh}
  \item Report bugs at \url{https://github.com/mdsumner/quadmesh/issues}
}

}
\author{
\strong{Maintainer}: Michael D. Sumner \email{mdsumner@gmail.com}

}
\keyword{internal}
