% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reproj.R
\name{reproj}
\alias{reproj}
\alias{reproj.quadmesh}
\title{Reprojection methods}
\usage{
\method{reproj}{quadmesh}(x, target, ..., source = NULL)
}
\arguments{
\item{x}{coordinates}

\item{target}{target specification (PROJ.4 string or epsg code)}

\item{...}{arguments passed to \code{\link{ptransform}}}

\item{source}{source specification (PROJ.4 string or epsg code)}
}
\description{
A `quadmesh` method for `reproj::reproj()`.
}
