% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/triangmesh.R
\name{triangmesh}
\alias{triangmesh}
\alias{triangmesh.matrix}
\alias{triangmesh.BasicRaster}
\title{Triangle mesh.}
\usage{
triangmesh(x, z = x, na.rm = FALSE, ..., texture = NULL,
  texture_filename = NULL)

\method{triangmesh}{matrix}(x, z = x, na.rm = FALSE, ...,
  texture = NULL, texture_filename = NULL)

\method{triangmesh}{BasicRaster}(x, z = x, na.rm = FALSE, ...,
  texture = NULL, texture_filename = NULL)
}
\arguments{
\item{x}{raster object for mesh structure}

\item{z}{raster object for height values}

\item{na.rm}{remove quads where missing values?}

\item{...}{ignored}

\item{texture}{optional input RGB raster, 3-layers}

\item{texture_filename}{optional input file path for PNG texture}
}
\value{
mesh3d (primitivetype triangle)
}
\description{
Triangle mesh.
}
\examples{
library(raster)
r <- setExtent(raster(volcano), extent(0, nrow(volcano), 0, ncol(volcano)))
tm <- triangmesh(r)
#rgl::shade3d(tm)
}
