% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use_crs.R
\name{use_crs}
\alias{use_crs}
\title{In-use coordinate system}
\usage{
use_crs(crs = NULL)
}
\arguments{
\item{crs}{provide PROJ string to set the value}
}
\value{
the current value of the in-use coordinate system (may be \code{NULL}), if
\code{crs} is provided it is returned after being set as the in-use value
}
\description{
Set or return the coordinate system currently in use.
}
\details{
If argument \code{crs} is NULL, the function returns the current value (which may be \code{NULL}).
}
\examples{
\dontrun{
use_crs()
use_crs("+proj=laea +datum=WGS84")
use_crs()
}
}
