% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quadmesh-package.R
\docType{data}
\name{xymap}
\alias{xymap}
\title{World map}
\format{
An object of class \code{matrix} (inherits from \code{array}) with 82403 rows and 2 columns.
}
\usage{
xymap
}
\description{
The world coastline coordinates. A simple matrix of lon, lat, separated by NA.
}
\details{
From the maps package, see 'data-raw/xymap.R'.
}
\keyword{datasets}
