% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quadmesh-package.R
\docType{package}
\name{quadmesh-package}
\alias{quadmesh-package}
\alias{_PACKAGE}
\title{quadmesh: Quadrangle Mesh}
\description{
\if{html}{\figure{logo.png}{options: align='right'}}

Create a standard quadrangle mesh from 'raster' data for flexible
 plotting and conversion to other mesh forms. The standard output 'quadmesh' is an 'mesh3d' object
 used by the 'rgl' package. This is used for plotting raster data in 3D with or without textures, and
 allows the application of a map projection without data loss.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/hypertidy/quadmesh}
  \item Report bugs at \url{https://github.com/hypertidy/quadmesh/issues}
}

}
\author{
\strong{Maintainer}: Michael D. Sumner \email{mdsumner@gmail.com}

}
\keyword{internal}
