% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quadmesh-package.R
\docType{package}
\name{quadmesh-package}
\alias{quadmesh-package}
\alias{_PACKAGE}
\title{quadmesh: Quadrangle Mesh}
\description{
\if{html}{\figure{logo.png}{options: align='right'}}

Create surface forms from matrix or 'raster' data for flexible plotting and
 conversion to other mesh types. The functions 'quadmesh' or 'triangmesh'
 produce a continuous surface as a 'mesh3d' object as used by the 'rgl'
 package. This is used for plotting raster data in 3D (optionally with
 texture), and allows the application of a map projection without data loss and 
 many processing applications that are restricted by inflexible regular grid rasters.
 There are discrete forms of these continuous surfaces available with
 'dquadmesh' and 'dtriangmesh' functions.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/hypertidy/quadmesh}
  \item Report bugs at \url{https://github.com/hypertidy/quadmesh/issues}
}

}
\author{
\strong{Maintainer}: Michael D. Sumner \email{mdsumner@gmail.com}

}
\keyword{internal}
