% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quadmesh-package.R
\docType{package}
\name{quadmesh-package}
\alias{quadmesh}
\alias{quadmesh-package}
\title{quadmesh: Quadrangle Mesh}
\description{
Gridded data need not be regular, quadmesh creates a mesh from rasters. Future versions may be more general.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/mdsumner/quadmesh}
  \item Report bugs at \url{https://github.com/mdsumner/quadmesh/issues}
}

}
\author{
\strong{Maintainer}: Michael D. Sumner \email{mdsumner@gmail.com}

}
\keyword{internal}
