% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/primitives.R
\name{triangulate_quads}
\alias{triangulate_quads}
\title{Triangles from quads}
\usage{
triangulate_quads(quad_index)
}
\arguments{
\item{quad_index}{}
}
\value{
matrix of triangle indices
}
\description{
Convert quad index to triangles.
}
\details{
Triangle pairs from each quad are interleaved in the result, so that neighbour
triangles from a single quad are together.
}
